Add files via upload
This is an ensemble simulation of the JWST's NIRCam Deep Field Simulation. This must be run in Google Collaboratory. The ensemble is currently set to the F277M filter, but this can be changed to any filter published in the NIRCam's documentation. The ensemble is currently set to run for 10 perturbations with a single run of each perturbation, but this can be easily changed in the ensemble control panel to whatever size ensemble is desired.