Add files via upload
This is an ensemble simulation of the JWST NIRCam deep field image. This must be run in google collaborator. The ensemble is currently set up with the F277W filter, however, this can be easily changed in the control panel to any NIRCam filter described in the NIRCam documentation. The ensemble is currently set up for 10 perturbations of the initial conditions with one run per perturbation, but this can be easily changed in the ensemble control panel to the desired number of perturbations and runs per perturbation.